refactor: 패키지 이동 경로 변경
