+++ src/main/java/com/depromeet/domain/auth/application/JwtTokenService.java
@@ -6,14 +6,17 @@ import com.depromeet.domain.auth.dto.AccessToken;
 import com.depromeet.domain.member.dao.MemberRepository;
 import com.depromeet.domain.member.domain.Member;
 import com.depromeet.domain.member.domain.MemberRole;
-import com.depromeet.global.config.security.PrincipalDetails;
 import com.depromeet.global.security.JwtTokenProvider;
-import java.util.NoSuchElementException;
+import com.depromeet.global.security.PrincipalDetails;
+
 import lombok.RequiredArgsConstructor;
+
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+
+import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
