+++ src/main/java/com/depromeet/global/util/SecurityUtil.java
@@ -1,6 +1,7 @@
 package com.depromeet.global.util;
 
-import com.depromeet.global.config.security.PrincipalDetails;
+import com.depromeet.global.security.PrincipalDetails;
+
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
